polytech/requirements.txt
PyQt==5.15.2
openai==1.68.2
